+++ griptape/artifacts/__init__.py
@@ -1,32 +1,39 @@
 from .base_artifact import BaseArtifact
+
+from .base_system_artifact import BaseSystemArtifact
 from .error_artifact import ErrorArtifact
 from .info_artifact import InfoArtifact
+
 from .text_artifact import TextArtifact
 from .json_artifact import JsonArtifact
-from .blob_artifact import BlobArtifact
-from .boolean_artifact import BooleanArtifact
 from .csv_row_artifact import CsvRowArtifact
+from .table_artifact import TableArtifact
+
 from .list_artifact import ListArtifact
-from .media_artifact import MediaArtifact
+
+from .blob_artifact import BlobArtifact
+
 from .image_artifact import ImageArtifact
 from .audio_artifact import AudioArtifact
+
 from .action_artifact import ActionArtifact
+
 from .generic_artifact import GenericArtifact
 
 
 __all__ = [
     "BaseArtifact",
+    "BaseSystemArtifact",
     "ErrorArtifact",
     "InfoArtifact",
     "TextArtifact",
     "JsonArtifact",
     "BlobArtifact",
-    "BooleanArtifact",
     "CsvRowArtifact",
     "ListArtifact",
-    "MediaArtifact",
     "ImageArtifact",
     "AudioArtifact",
     "ActionArtifact",
     "GenericArtifact",
+    "TableArtifact",
 ]