+++ griptape/artifacts/action_artifact.py
@@ -5,15 +5,14 @@ from typing import TYPE_CHECKING
 from attrs import define, field
 
 from griptape.artifacts import BaseArtifact
-from griptape.mixins import SerializableMixin
 
 if TYPE_CHECKING:
     from griptape.common import ToolAction
 
 
 @define()
-class ActionArtifact(BaseArtifact, SerializableMixin):
+class ActionArtifact(BaseArtifact):
     value: ToolAction = field(metadata={"serializable": True})
 
-    def __add__(self, other: BaseArtifact) -> ActionArtifact:
-        raise NotImplementedError
+    def to_text(self) -> str:
+        return str(self.value)