+++ griptape/artifacts/audio_artifact.py
@@ -1,12 +1,20 @@
 from __future__ import annotations
 
-from attrs import define
+from attrs import define, field
 
-from griptape.artifacts import MediaArtifact
+from griptape.artifacts import BaseArtifact
 
 
 @define
-class AudioArtifact(MediaArtifact):
-    """AudioArtifact is a type of MediaArtifact representing audio."""
+class AudioArtifact(BaseArtifact):
+    """AudioArtifact is a type of Artifact representing audio."""
 
-    media_type: str = "audio"
+    value: bytes = field(metadata={"serializable": True})
+    format: str = field(kw_only=True, metadata={"serializable": True})
+
+    @property
+    def mime_type(self) -> str:
+        return f"audio/{self.format}"
+
+    def to_text(self) -> str:
+        return f"Audio, format: {self.format}, size: {len(self.value)} bytes"