+++ griptape/artifacts/csv_row_artifact.py
@@ -2,22 +2,29 @@ from __future__ import annotations
 
 import csv
 import io
+import json
+from typing import Any
 
-from attrs import define, field
+from attrs import Converter, define, field
 
-from griptape.artifacts import BaseArtifact, TextArtifact
+from griptape.artifacts import TextArtifact
 
 
 @define
 class CsvRowArtifact(TextArtifact):
-    value: dict[str, str] = field(converter=BaseArtifact.value_to_dict, metadata={"serializable": True})
+    value: dict[str, str] = field(
+        converter=Converter(lambda value: CsvRowArtifact.value_to_dict(value)), metadata={"serializable": True}
+    )
     delimiter: str = field(default=",", kw_only=True, metadata={"serializable": True})
-
-    def __add__(self, other: BaseArtifact) -> CsvRowArtifact:
-        return CsvRowArtifact(self.value | other.value)
 
     def __bool__(self) -> bool:
         return len(self) > 0
+
+    @classmethod
+    def value_to_dict(cls, value: Any) -> dict:
+        dict_value = value if isinstance(value, dict) else json.loads(value)
+
+        return dict(dict_value.items())
 
     def to_text(self) -> str:
         with io.StringIO() as csvfile: