+++ griptape/artifacts/error_artifact.py
@@ -4,13 +4,10 @@ from typing import Optional
 
 from attrs import define, field
 
-from griptape.artifacts import BaseArtifact
+from griptape.artifacts import BaseSystemArtifact
 
 
 @define
-class ErrorArtifact(BaseArtifact):
+class ErrorArtifact(BaseSystemArtifact):
     value: str = field(converter=str, metadata={"serializable": True})
     exception: Optional[Exception] = field(default=None, kw_only=True, metadata={"serializable": False})
-
-    def __add__(self, other: BaseArtifact) -> ErrorArtifact:
-        return ErrorArtifact(self.value + other.value)