+++ griptape/loaders/csv_loader.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Optional, cast
 
 from attrs import define, field
 
-from griptape.artifacts import CsvRowArtifact
+from griptape.artifacts import TableArtifact
 from griptape.loaders import BaseLoader
 
 if TYPE_CHECKING:
@@ -19,33 +19,28 @@ class CsvLoader(BaseLoader):
     delimiter: str = field(default=",", kw_only=True)
     encoding: str = field(default="utf-8", kw_only=True)
 
-    def load(self, source: bytes | str, *args, **kwargs) -> list[CsvRowArtifact]:
-        artifacts = []
-
+    def load(self, source: bytes | str, *args, **kwargs) -> TableArtifact:
         if isinstance(source, bytes):
             source = source.decode(encoding=self.encoding)
         elif isinstance(source, (bytearray, memoryview)):
             raise ValueError(f"Unsupported source type: {type(source)}")
 
         reader = csv.DictReader(StringIO(source), delimiter=self.delimiter)
-        chunks = [CsvRowArtifact(row) for row in reader]
+
+        artifact = TableArtifact(list(reader), delimiter=self.delimiter, fieldnames=reader.fieldnames)
 
         if self.embedding_driver:
-            for chunk in chunks:
-                chunk.generate_embedding(self.embedding_driver)
+            artifact.generate_embedding(self.embedding_driver)
 
-        for chunk in chunks:
-            artifacts.append(chunk)
-
-        return artifacts
+        return artifact
 
     def load_collection(
         self,
         sources: list[bytes | str],
         *args,
         **kwargs,
-    ) -> dict[str, list[CsvRowArtifact]]:
+    ) -> dict[str, TableArtifact]:
         return cast(
-            dict[str, list[CsvRowArtifact]],
+            dict[str, TableArtifact],
             super().load_collection(sources, *args, **kwargs),
         )