+++ griptape/loaders/dataframe_loader.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING, Optional, cast
 
 from attrs import define, field
 
-from griptape.artifacts import CsvRowArtifact
+from griptape.artifacts import TableArtifact
 from griptape.loaders import BaseLoader
 from griptape.utils import import_optional_dependency
 from griptape.utils.hash import str_to_hash
@@ -19,22 +19,16 @@ if TYPE_CHECKING:
 class DataFrameLoader(BaseLoader):
     embedding_driver: Optional[BaseEmbeddingDriver] = field(default=None, kw_only=True)
 
-    def load(self, source: DataFrame, *args, **kwargs) -> list[CsvRowArtifact]:
-        artifacts = []
-
-        chunks = [CsvRowArtifact(row) for row in source.to_dict(orient="records")]
+    def load(self, source: DataFrame, *args, **kwargs) -> TableArtifact:
+        artifact = TableArtifact(list(source.to_dict(orient="records")))
 
         if self.embedding_driver:
-            for chunk in chunks:
-                chunk.generate_embedding(self.embedding_driver)
+            artifact.generate_embedding(self.embedding_driver)
 
-        for chunk in chunks:
-            artifacts.append(chunk)
+        return artifact
 
-        return artifacts
-
-    def load_collection(self, sources: list[DataFrame], *args, **kwargs) -> dict[str, list[CsvRowArtifact]]:
-        return cast(dict[str, list[CsvRowArtifact]], super().load_collection(sources, *args, **kwargs))
+    def load_collection(self, sources: list[DataFrame], *args, **kwargs) -> dict[str, TableArtifact]:
+        return cast(dict[str, TableArtifact], super().load_collection(sources, *args, **kwargs))
 
     def to_key(self, source: DataFrame, *args, **kwargs) -> str:
         hash_pandas_object = import_optional_dependency("pandas.core.util.hashing").hash_pandas_object