+++ griptape/mixins/media_artifact_file_output_mixin.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Optional
 from attrs import Attribute, define, field
 
 if TYPE_CHECKING:
-    from griptape.artifacts import BlobArtifact
+    from griptape.artifacts import BaseArtifact
 
 
 @define(slots=False)
@@ -31,7 +31,7 @@ class BlobArtifactFileOutputMixin:
         if self.output_dir:
             raise ValueError("Can't have both output_dir and output_file specified.")
 
-    def _write_to_file(self, artifact: BlobArtifact) -> None:
+    def _write_to_file(self, artifact: BaseArtifact) -> None:
         if self.output_file:
             outfile = self.output_file
         elif self.output_dir:
@@ -42,4 +42,4 @@ class BlobArtifactFileOutputMixin:
         if os.path.dirname(outfile):
             os.makedirs(os.path.dirname(outfile), exist_ok=True)
 
-        Path(outfile).write_bytes(artifact.value)
+        Path(outfile).write_bytes(artifact.to_bytes())