+++ tests/unit/artifacts/test_csv_row_artifact.py
@@ -7,12 +7,6 @@ class TestCsvRowArtifact:
         assert CsvRowArtifact({"foo": {"bar": "baz"}}).value == {"foo": {"bar": "baz"}}
         assert CsvRowArtifact('{"foo": "bar"}').value == {"foo": "bar"}
 
-    def test___add__(self):
-        assert (CsvRowArtifact({"test1": "foo"}) + CsvRowArtifact({"test2": "bar"})).value == {
-            "test1": "foo",
-            "test2": "bar",
-        }
-
     def test_to_text(self):
         assert CsvRowArtifact({"test1": "foo|bar", "test2": 1}, delimiter="|").to_text() == 'test1|test2\r\n"foo|bar"|1'