+++ tests/unit/artifacts/test_json_artifact.py
@@ -1,8 +1,6 @@
 import json
 
-import pytest
-
-from griptape.artifacts import JsonArtifact, TextArtifact
+from griptape.artifacts import JsonArtifact
 
 
 class TestJsonArtifact:
@@ -15,10 +13,6 @@ class TestJsonArtifact:
         assert JsonArtifact([{"foo": {"bar": "baz"}}]).value == json.loads(json.dumps([{"foo": {"bar": "baz"}}]))
         assert JsonArtifact(None).value == json.loads(json.dumps(None))
         assert JsonArtifact("foo").value == json.loads(json.dumps("foo"))
-
-    def test___add__(self):
-        with pytest.raises(NotImplementedError):
-            JsonArtifact({"foo": "bar"}) + TextArtifact("invalid json")
 
     def test_to_text(self):
         assert JsonArtifact({"foo": "bar"}).to_text() == json.dumps({"foo": "bar"})