+++ tests/unit/loaders/test_sql_loader.py
@@ -35,14 +35,14 @@ class TestSqlLoader:
         return sql_loader
 
     def test_load(self, loader):
-        artifacts = loader.load("SELECT * FROM test_table;")
+        artifact = loader.load("SELECT * FROM test_table;")
 
-        assert len(artifacts) == 3
-        assert artifacts[0].value == {"id": 1, "name": "Alice", "age": 25, "city": "New York"}
-        assert artifacts[1].value == {"id": 2, "name": "Bob", "age": 30, "city": "Los Angeles"}
-        assert artifacts[2].value == {"id": 3, "name": "Charlie", "age": 22, "city": "Chicago"}
+        assert len(artifact) == 3
+        assert artifact.value[0] == {"id": 1, "name": "Alice", "age": 25, "city": "New York"}
+        assert artifact.value[1] == {"id": 2, "name": "Bob", "age": 30, "city": "Los Angeles"}
+        assert artifact.value[2] == {"id": 3, "name": "Charlie", "age": 22, "city": "Chicago"}
 
-        assert artifacts[0].embedding == [0, 1]
+        assert artifact.embedding == [0, 1]
 
     def test_load_collection(self, loader):
         artifacts = loader.load_collection(["SELECT * FROM test_table LIMIT 1;", "SELECT * FROM test_table LIMIT 2;"])
@@ -52,10 +52,10 @@ class TestSqlLoader:
             loader.to_key("SELECT * FROM test_table LIMIT 2;"),
         ]
 
-        assert [a.value for artifact_list in artifacts.values() for a in artifact_list] == [
+        assert [a for artifact_table in artifacts.values() for a in artifact_table.value] == [
             {"age": 25, "city": "New York", "id": 1, "name": "Alice"},
             {"age": 25, "city": "New York", "id": 1, "name": "Alice"},
             {"age": 30, "city": "Los Angeles", "id": 2, "name": "Bob"},
         ]
 
-        assert list(artifacts.values())[0][0].embedding == [0, 1]
+        assert list(artifacts.values())[0].embedding == [0, 1]